Tests/LibWeb/Text/expected/kerning-across-bidi-neutral.txt

Wrapping and non-wrapping text have equal widths: true
